Orca: Upgrade tutorial xshards to tf2 (#7916)
* ci(notebook-xshards-py38): add

* ci(xshards-image-nightly): add

* test(nightly_test.yml): trigger

* fix(action.yml): add "step per epoch"

* fix(ipynb): change api to tf2

* fix(tabular): use "int" instead of np.long

* fix(action.yml): change py37 test

* perf(notebook): calc train steps

* fix: use "len()"

* fix: use math.ceil

* fix(ipynb): revert metadata changes

* ci(PR_validation.yml): revert

* test:reinit workflow

* fix: remove \n

* fix auto_mpg \n

* fix: revert py env name

* fix(xshards-py37): upgrade

* fix(diabetes.py): fix style

* style: import math at front

+++ python/orca/tutorial/xshards/tabular_playground_series.py
@@ -71,7 +71,7 @@ val_shard = scale.transform(val_shard)
 # Change data types
 def change_data_type(df):
     df['x_scaled'] = df['x_scaled'].apply(lambda x: np.array(x, dtype=np.float32))
-    df['target'] = df['target'].apply(lambda x: np.long(x))
+    df['target'] = df['target'].apply(lambda x: int(x))
     return df
 train_shard = train_shard.transform_shard(change_data_type)
 val_shard = val_shard.transform_shard(change_data_type)
